Splittet kartpakke-modalen opp i flere filer
