+++ src/app/pages/offline-map/offline-package-modal/offline-package-modal.component.html
@@ -1,0 +1,64 @@
+<ion-header translucent>
+  <ion-toolbar>
+    <ion-title>Kartpakke {{packageOnServer.properties.xyz[0]}}-{{packageOnServer.properties.xyz[1]}}-{{packageOnServer.properties.xyz[2]}}</ion-title>
+    <ion-buttons slot="end">
+      <ion-button (click)="dismiss()">Lukk</ion-button>
+    </ion-buttons>
+  </ion-toolbar>
+</ion-header>
+<ion-content>
+  <div class="map-container">
+    <app-map
+      [showMapControls]="false"
+      [showUserLocation]="false"
+      [showSupportMaps]="false"
+      [center]="center"
+      [zoom]="zoom"
+      (mapReady)="showTileOnMap($event)"
+    ></app-map>
+  </div>
+  <ion-grid>
+    <ion-row>
+      <ion-col size="4">Produsert:</ion-col>
+      <ion-col>{{ packageOnServer.properties.lastModified }}</ion-col>
+    </ion-row>
+    <ion-row>
+      <ion-col size="4">Størrelse:</ion-col>
+      <ion-col>{{ packageOnServer.properties.sizeInMb }} MB</ion-col>
+    </ion-row>
+    <div *ngIf="downloadedPackage$ | async as downloadedPackage">
+      <ion-row>
+        <ion-col size="4">Status:</ion-col>
+        <ion-col>
+          {{ downloadedPackage.progress?.description }}
+          <span *ngIf="!!downloadedPackage.downloadComplete">Lastet ned {{ downloadedPackage.downloadComplete | formatDate:true:false:true | async }} </span>
+          <span *ngIf="!downloadedPackage.downloadComplete">({{getPercentage(downloadedPackage) +'%' }})</span>
+        </ion-col>
+        <ion-col size="1">                  
+          <ion-icon *ngIf="!!downloadedPackage.downloadComplete" name="checkmark"></ion-icon>
+          <ion-icon *ngIf="downloadedPackage.progress?.step === 0" name="cloud-download-outline"></ion-icon>
+          <ion-icon *ngIf="downloadedPackage.progress?.step === 1" name="folder-open-outline"></ion-icon>
+        </ion-col>
+      </ion-row>              
+      <ion-row *ngIf="!!downloadedPackage.downloadComplete">
+        <ion-col>
+          <ion-button (click)="delete()" expand="block" color="danger">Slett</ion-button>
+        </ion-col>
+      </ion-row>
+      <ion-row *ngIf="!downloadedPackage.downloadComplete">
+        <ion-col>
+          <ion-button expand="block" color="danger">Avbryt nedlasting (TODO)</ion-button>
+        </ion-col>
+      </ion-row>
+    </div>
+    <ion-row *ngIf="!(downloadedPackage$ | async)">
+      <ion-col>
+        <ion-button
+          (click)="startDownload()"
+          expand="block"
+          color="varsom-orange"
+        >Last ned</ion-button>
+      </ion-col>
+    </ion-row>
+  </ion-grid>
+</ion-content>
